// src/actions/create-shipping-address/index.ts
"use server";

import { eq } from "drizzle-orm";
import { revalidatePath } from "next/cache";
import { headers } from "next/headers";
import type z from "zod";

import { db } from "@/db";
import { cartTable, shippingAddressTable } from "@/db/schema";
import { auth } from "@/lib/auth";

import { createShippingAddressSchema } from "./schema";

export const createShippingAddress = async (
  data: z.infer<typeof createShippingAddressSchema>,
) => {
  createShippingAddressSchema.parse(data);

  const session = await auth.api.getSession({
    headers: await headers(),
  });

  if (!session?.user) {
    throw new Error("Unauthorized");
  }

  const [shippingAddress] = await db
    .insert(shippingAddressTable)
    .values({
      userId: session.user.id,
      recipientName: data.recipientName,
      street: data.street,
      number: data.number,
      complement: data.complement,
      city: data.city,
      state: data.state,
      neighborhood: data.neighborhood,
      zipCode: data.zipCode,
      country: "Brasil",
      phone: data.phone,
      email: data.email,
      cpfOrCnpj: data.cpfOrCnpj,
    })
    .returning();

  const cart = await db.query.cartTable.findFirst({
    where: eq(cartTable.userId, session.user.id),
  });

  if (cart) {
    await db
      .update(cartTable)
      .set({
        shippingAddressId: shippingAddress.id,
      })
      .where(eq(cartTable.id, cart.id));
  }

  revalidatePath("/cart/identification");

  return shippingAddress;
};
